WebServiceKodlar/requirements.txt
cv2
Flask
gevent
gunicorn
keras
numpy
Image
h5py
pillow
tensorflow
werkzeug